create resources package
