+++ internal/k8s/resources/resources_test.go
@@ -1,0 +1,397 @@
+package resources
+
+import (
+	"reflect"
+	"testing"
+
+	"github.com/nginxinc/kubernetes-ingress/internal/configs"
+	v1 "k8s.io/api/core/v1"
+	meta_v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/util/intstr"
+)
+
+func TestFindProbeForPods(t *testing.T) {
+	pods := []*v1.Pod{
+		{
+			Spec: v1.PodSpec{
+				Containers: []v1.Container{
+					{
+						ReadinessProbe: &v1.Probe{
+							Handler: v1.Handler{
+								HTTPGet: &v1.HTTPGetAction{
+									Path: "/",
+									Host: "asdf.com",
+									Port: intstr.IntOrString{
+										IntVal: 80,
+									},
+								},
+							},
+							PeriodSeconds: 42,
+						},
+						Ports: []v1.ContainerPort{
+							{
+								Name:          "name",
+								ContainerPort: 80,
+								Protocol:      v1.ProtocolTCP,
+								HostIP:        "1.2.3.4",
+							},
+						},
+					},
+				},
+			},
+		},
+	}
+	svcPort := v1.ServicePort{
+		TargetPort: intstr.FromInt(80),
+	}
+	probe := findProbeForPods(pods, &svcPort)
+	if probe == nil || probe.PeriodSeconds != 42 {
+		t.Errorf("ServicePort.TargetPort as int match failed: %+v", probe)
+	}
+
+	svcPort = v1.ServicePort{
+		TargetPort: intstr.FromString("name"),
+		Protocol:   v1.ProtocolTCP,
+	}
+	probe = findProbeForPods(pods, &svcPort)
+	if probe == nil || probe.PeriodSeconds != 42 {
+		t.Errorf("ServicePort.TargetPort as string failed: %+v", probe)
+	}
+
+	svcPort = v1.ServicePort{
+		TargetPort: intstr.FromInt(80),
+		Protocol:   v1.ProtocolTCP,
+	}
+	probe = findProbeForPods(pods, &svcPort)
+	if probe == nil || probe.PeriodSeconds != 42 {
+		t.Errorf("ServicePort.TargetPort as int failed: %+v", probe)
+	}
+
+	svcPort = v1.ServicePort{
+		Port: 80,
+	}
+	probe = findProbeForPods(pods, &svcPort)
+	if probe == nil || probe.PeriodSeconds != 42 {
+		t.Errorf("ServicePort.Port should match if TargetPort is not set: %+v", probe)
+	}
+
+	svcPort = v1.ServicePort{
+		TargetPort: intstr.FromString("wrong_name"),
+	}
+	probe = findProbeForPods(pods, &svcPort)
+	if probe != nil {
+		t.Errorf("ServicePort.TargetPort should not have matched string: %+v", probe)
+	}
+
+	svcPort = v1.ServicePort{
+		TargetPort: intstr.FromInt(22),
+	}
+	probe = findProbeForPods(pods, &svcPort)
+	if probe != nil {
+		t.Errorf("ServicePort.TargetPort should not have matched int: %+v", probe)
+	}
+
+	svcPort = v1.ServicePort{
+		Port: 22,
+	}
+	probe = findProbeForPods(pods, &svcPort)
+	if probe != nil {
+		t.Errorf("ServicePort.Port mismatch: %+v", probe)
+	}
+}
+
+func TestGetPodOwnerTypeAndName(t *testing.T) {
+	tests := []struct {
+		desc    string
+		expType string
+		expName string
+		pod     *v1.Pod
+	}{
+		{
+			desc:    "deployment",
+			expType: "deployment",
+			expName: "deploy-name",
+			pod:     &v1.Pod{ObjectMeta: createTestObjMeta("Deployment", "deploy-name", true)},
+		},
+		{
+			desc:    "stateful set",
+			expType: "statefulset",
+			expName: "statefulset-name",
+			pod:     &v1.Pod{ObjectMeta: createTestObjMeta("StatefulSet", "statefulset-name", true)},
+		},
+		{
+			desc:    "daemon set",
+			expType: "daemonset",
+			expName: "daemonset-name",
+			pod:     &v1.Pod{ObjectMeta: createTestObjMeta("DaemonSet", "daemonset-name", true)},
+		},
+		{
+			desc:    "replica set with no pod hash",
+			expType: "deployment",
+			expName: "replicaset-name",
+			pod:     &v1.Pod{ObjectMeta: createTestObjMeta("ReplicaSet", "replicaset-name", false)},
+		},
+		{
+			desc:    "replica set with pod hash",
+			expType: "deployment",
+			expName: "replicaset-name",
+			pod: &v1.Pod{
+				ObjectMeta: createTestObjMeta("ReplicaSet", "replicaset-name-67c6f7c5fd", true),
+			},
+		},
+		{
+			desc:    "nil controller should use default values",
+			expType: "deployment",
+			expName: "deploy-name",
+			pod: &v1.Pod{
+				ObjectMeta: meta_v1.ObjectMeta{
+					OwnerReferences: []meta_v1.OwnerReference{
+						{
+							Name:       "deploy-name",
+							Controller: nil,
+						},
+					},
+				},
+			},
+		},
+	}
+
+	for _, test := range tests {
+		t.Run(test.desc, func(t *testing.T) {
+			actualType, actualName := getPodOwnerTypeAndName(test.pod)
+			if actualType != test.expType {
+				t.Errorf("getPodOwnerTypeAndName() returned %s for owner type but expected %s", actualType, test.expType)
+			}
+			if actualName != test.expName {
+				t.Errorf("getPodOwnerTypeAndName() returned %s for owner name but expected %s", actualName, test.expName)
+			}
+		})
+	}
+}
+
+func TestGetEndpointsBySubselectedPods(t *testing.T) {
+	boolPointer := func(b bool) *bool { return &b }
+	tests := []struct {
+		desc        string
+		targetPort  int32
+		svcEps      v1.Endpoints
+		expectedEps []PodEndpoint
+	}{
+		{
+			desc:       "find one endpoint",
+			targetPort: 80,
+			expectedEps: []PodEndpoint{
+				{
+					Address: "1.2.3.4:80",
+					MeshPodOwner: configs.MeshPodOwner{
+						OwnerType: "deployment",
+						OwnerName: "deploy-1",
+					},
+				},
+			},
+		},
+		{
+			desc:        "targetPort mismatch",
+			targetPort:  21,
+			expectedEps: nil,
+		},
+	}
+
+	pods := []*v1.Pod{
+		{
+			ObjectMeta: meta_v1.ObjectMeta{
+				OwnerReferences: []meta_v1.OwnerReference{
+					{
+						Kind:       "Deployment",
+						Name:       "deploy-1",
+						Controller: boolPointer(true),
+					},
+				},
+			},
+			Status: v1.PodStatus{
+				PodIP: "1.2.3.4",
+			},
+		},
+	}
+
+	svcEps := v1.Endpoints{
+		Subsets: []v1.EndpointSubset{
+			{
+				Addresses: []v1.EndpointAddress{
+					{
+						IP:       "1.2.3.4",
+						Hostname: "asdf.com",
+					},
+				},
+				Ports: []v1.EndpointPort{
+					{
+						Port: 80,
+					},
+				},
+			},
+		},
+	}
+
+	for _, test := range tests {
+		t.Run(test.desc, func(t *testing.T) {
+			gotEndps := getEndpointsBySubselectedPods(test.targetPort, pods, svcEps)
+			if !reflect.DeepEqual(gotEndps, test.expectedEps) {
+				t.Errorf("getEndpointsBySubselectedPods() = %v, want %v", gotEndps, test.expectedEps)
+			}
+		})
+	}
+}
+
+func createTestObjMeta(kind, name string, podHashLabel bool) meta_v1.ObjectMeta {
+	controller := true
+	meta := meta_v1.ObjectMeta{
+		OwnerReferences: []meta_v1.OwnerReference{
+			{
+				Kind:       kind,
+				Name:       name,
+				Controller: &controller,
+			},
+		},
+	}
+	if podHashLabel {
+		meta.Labels = map[string]string{
+			"pod-template-hash": "67c6f7c5fd",
+		}
+	}
+	return meta
+}
+
+func TestComparePorts(t *testing.T) {
+	scenarios := []struct {
+		sp       v1.ServicePort
+		cp       v1.ContainerPort
+		expected bool
+	}{
+		{
+			// match TargetPort.strval and Protocol
+			v1.ServicePort{
+				TargetPort: intstr.FromString("name"),
+				Protocol:   v1.ProtocolTCP,
+			},
+			v1.ContainerPort{
+				Name:          "name",
+				Protocol:      v1.ProtocolTCP,
+				ContainerPort: 80,
+			},
+			true,
+		},
+		{
+			// don't match Name and Protocol
+			v1.ServicePort{
+				Name:     "name",
+				Protocol: v1.ProtocolTCP,
+			},
+			v1.ContainerPort{
+				Name:          "name",
+				Protocol:      v1.ProtocolTCP,
+				ContainerPort: 80,
+			},
+			false,
+		},
+		{
+			// TargetPort intval mismatch, don't match by TargetPort.Name
+			v1.ServicePort{
+				Name:       "name",
+				TargetPort: intstr.FromInt(80),
+			},
+			v1.ContainerPort{
+				Name:          "name",
+				ContainerPort: 81,
+			},
+			false,
+		},
+		{
+			// match by TargetPort intval
+			v1.ServicePort{
+				TargetPort: intstr.IntOrString{
+					IntVal: 80,
+				},
+			},
+			v1.ContainerPort{
+				ContainerPort: 80,
+			},
+			true,
+		},
+		{
+			// Fall back on ServicePort.Port if TargetPort is empty
+			v1.ServicePort{
+				Name: "name",
+				Port: 80,
+			},
+			v1.ContainerPort{
+				Name:          "name",
+				ContainerPort: 80,
+			},
+			true,
+		},
+		{
+			// TargetPort intval mismatch
+			v1.ServicePort{
+				TargetPort: intstr.FromInt(80),
+			},
+			v1.ContainerPort{
+				ContainerPort: 81,
+			},
+			false,
+		},
+		{
+			// don't match empty ports
+			v1.ServicePort{},
+			v1.ContainerPort{},
+			false,
+		},
+	}
+
+	for _, scen := range scenarios {
+		if scen.expected != compareContainerPortAndServicePort(scen.cp, scen.sp) {
+			t.Errorf("Expected: %v, ContainerPort: %v, ServicePort: %v", scen.expected, scen.cp, scen.sp)
+		}
+	}
+}
+
+func TestGetServicePortForIngressPort(t *testing.T) {
+	svc := v1.Service{
+		TypeMeta: meta_v1.TypeMeta{},
+		ObjectMeta: meta_v1.ObjectMeta{
+			Name:      "coffee-svc",
+			Namespace: "default",
+		},
+		Spec: v1.ServiceSpec{
+			Ports: []v1.ServicePort{
+				{
+					Name:       "foo",
+					Port:       80,
+					TargetPort: intstr.FromInt(22),
+				},
+			},
+		},
+		Status: v1.ServiceStatus{},
+	}
+	ingSvcPort := intstr.FromString("foo")
+	svcPort := getServicePortForIngressPort(ingSvcPort, &svc)
+	if svcPort == nil || svcPort.Port != 80 {
+		t.Errorf("TargetPort string match failed: %+v", svcPort)
+	}
+
+	ingSvcPort = intstr.FromInt(80)
+	svcPort = getServicePortForIngressPort(ingSvcPort, &svc)
+	if svcPort == nil || svcPort.Port != 80 {
+		t.Errorf("TargetPort int match failed: %+v", svcPort)
+	}
+
+	ingSvcPort = intstr.FromInt(22)
+	svcPort = getServicePortForIngressPort(ingSvcPort, &svc)
+	if svcPort != nil {
+		t.Errorf("Mismatched ints should not return port: %+v", svcPort)
+	}
+	ingSvcPort = intstr.FromString("bar")
+	svcPort = getServicePortForIngressPort(ingSvcPort, &svc)
+	if svcPort != nil {
+		t.Errorf("Mismatched strings should not return port: %+v", svcPort)
+	}
+}
