+++ internal/k8s/utils.go
@@ -22,135 +22,10 @@ import (
 	"strings"
 
 	"github.com/golang/glog"
-	v1 "k8s.io/api/core/v1"
 	networking "k8s.io/api/networking/v1beta1"
-
-	"k8s.io/apimachinery/pkg/labels"
-	"k8s.io/apimachinery/pkg/util/intstr"
 	"k8s.io/apimachinery/pkg/util/version"
 	"k8s.io/client-go/kubernetes"
-	"k8s.io/client-go/tools/cache"
 )
-
-// storeToIngressLister makes a Store that lists Ingress.
-// TODO: Move this to cache/listers post 1.1.
-type storeToIngressLister struct {
-	cache.Store
-}
-
-// GetByKeySafe calls Store.GetByKeySafe and returns a copy of the ingress so it is
-// safe to modify.
-func (s *storeToIngressLister) GetByKeySafe(key string) (ing *networking.Ingress, exists bool, err error) {
-	item, exists, err := s.Store.GetByKey(key)
-	if !exists || err != nil {
-		return nil, exists, err
-	}
-	ing = item.(*networking.Ingress).DeepCopy()
-	return
-}
-
-// List lists all Ingress' in the store.
-func (s *storeToIngressLister) List() (ing networking.IngressList, err error) {
-	for _, m := range s.Store.List() {
-		ing.Items = append(ing.Items, *(m.(*networking.Ingress)).DeepCopy())
-	}
-	return ing, nil
-}
-
-// GetServiceIngress gets all the Ingress' that have rules pointing to a service.
-// Note that this ignores services without the right nodePorts.
-func (s *storeToIngressLister) GetServiceIngress(svc *v1.Service) (ings []networking.Ingress, err error) {
-	for _, m := range s.Store.List() {
-		ing := *m.(*networking.Ingress).DeepCopy()
-		if ing.Namespace != svc.Namespace {
-			continue
-		}
-		if ing.Spec.Backend != nil {
-			if ing.Spec.Backend.ServiceName == svc.Name {
-				ings = append(ings, ing)
-			}
-		}
-		for _, rules := range ing.Spec.Rules {
-			if rules.IngressRuleValue.HTTP == nil {
-				continue
-			}
-			for _, p := range rules.IngressRuleValue.HTTP.Paths {
-				if p.Backend.ServiceName == svc.Name {
-					ings = append(ings, ing)
-				}
-			}
-		}
-	}
-	if len(ings) == 0 {
-		err = fmt.Errorf("No ingress for service %v", svc.Name)
-	}
-	return
-}
-
-// storeToConfigMapLister makes a Store that lists ConfigMaps
-type storeToConfigMapLister struct {
-	cache.Store
-}
-
-// List lists all Ingress' in the store.
-func (s *storeToConfigMapLister) List() (cfgm v1.ConfigMapList, err error) {
-	for _, m := range s.Store.List() {
-		cfgm.Items = append(cfgm.Items, *(m.(*v1.ConfigMap)))
-	}
-	return cfgm, nil
-}
-
-// indexerToPodLister makes a Indexer that lists Pods.
-type indexerToPodLister struct {
-	cache.Indexer
-}
-
-// ListByNamespace lists all Pods in the indexer for a given namespace that match the provided selector.
-func (ipl indexerToPodLister) ListByNamespace(ns string, selector labels.Selector) (pods []*v1.Pod, err error) {
-	err = cache.ListAllByNamespace(ipl.Indexer, ns, selector, func(m interface{}) {
-		pods = append(pods, m.(*v1.Pod))
-	})
-	return pods, err
-}
-
-// storeToEndpointLister makes a Store that lists Endponts
-type storeToEndpointLister struct {
-	cache.Store
-}
-
-// GetServiceEndpoints returns the endpoints of a service, matched on service name.
-func (s *storeToEndpointLister) GetServiceEndpoints(svc *v1.Service) (ep v1.Endpoints, err error) {
-	for _, m := range s.Store.List() {
-		ep = *m.(*v1.Endpoints)
-		if svc.Name == ep.Name && svc.Namespace == ep.Namespace {
-			return ep, nil
-		}
-	}
-	return ep, fmt.Errorf("could not find endpoints for service: %v", svc.Name)
-}
-
-// findPort locates the container port for the given pod and portName.  If the
-// targetPort is a number, use that.  If the targetPort is a string, look that
-// string up in all named ports in all containers in the target pod.  If no
-// match is found, fail.
-func findPort(pod *v1.Pod, svcPort v1.ServicePort) (int32, error) {
-	portName := svcPort.TargetPort
-	switch portName.Type {
-	case intstr.String:
-		name := portName.StrVal
-		for _, container := range pod.Spec.Containers {
-			for _, port := range container.Ports {
-				if port.Name == name && port.Protocol == svcPort.Protocol {
-					return port.ContainerPort, nil
-				}
-			}
-		}
-	case intstr.Int:
-		return int32(portName.IntValue()), nil
-	}
-
-	return 0, fmt.Errorf("no suitable port for manifest: %s", pod.UID)
-}
 
 // isMinion determines is an ingress is a minion or not
 func isMinion(ing *networking.Ingress) bool {
